+++ history.md
@@ -15,3 +15,4 @@
     * Fix: query not being remove from urls
     * Change: make middleware functions not use a path
     * Fix: A route would not be used if it ended with '/' because it was always removed, this issue is fixed
+    * Add: support for byte ranges on static content